Atualizando dados de questionário socioeconômico

+++ humv_backend/src/main/java/br/edu/ufrb/lasis/humv/dao/QuestionarioSocioeconomicoDAO.java
@@ -45,7 +45,7 @@ public class QuestionarioSocioeconomicoDAO extends GenericDAO<QuestionarioSocioe
 	}
 
 	@Transactional
-	public QuestionarioSocioeconomico findByKey(Integer id) {
+	public QuestionarioSocioeconomico findByKey(BigInteger id) {
 		return (QuestionarioSocioeconomico) getCriteria().add(Restrictions.eq("id", id)).uniqueResult();
 	}
